Use long instead of NSInteger to get rid of build warning.
As NSInteger is a typedef of long, this change makes no difference.

+++ XWalkView/XWalkView/XWalkWebView.m
@@ -103,7 +103,7 @@
         return nil;
     }
 
-    NSInteger port = 8080;
+    long port = 8080;
     if (![XWalkView httpServer].isRunning) {
         [[XWalkView httpServer] addGETHandlerForBasePath:@"/" directoryPath:readAccessURL.path indexFilename:nil cacheAge:3600 allowRangeRequests:YES];
         [[XWalkView httpServer] startWithPort:port bonjourName:nil];
